refactor: deduplicate frontend types and utilities
- Remove duplicate MetricsState, AuthState, LogsState interfaces from
  types/index.ts (stores are the SSOT)
- Extract formatUptime() and formatNumber() into web/src/utils/format.ts
- Unify two different formatUptime implementations (Overview.tsx showed
  "3d 2h", System.tsx showed "3d 2h 45m") into single robust version

Closes #83, closes #84
