+++ web/src/pages/Overview.tsx
@@ -1,6 +1,7 @@
 import { useEffect } from 'react';
 import { useMetricsStore } from '../stores/metricsStore';
 import MetricCard from '../components/MetricCard';
+import { formatUptime, formatNumber } from '../utils/format';
 import {
   Activity,
   AlertTriangle,
@@ -27,21 +28,6 @@ export default function Overview() {
   useEffect(() => {
     fetchStats();
   }, [fetchStats]);
-
-  const formatNumber = (n: number): string => {
-    if (n >= 1_000_000) return `${(n / 1_000_000).toFixed(1)}M`;
-    if (n >= 1_000) return `${(n / 1_000).toFixed(1)}K`;
-    return String(n);
-  };
-
-  const formatUptime = (seconds: number): string => {
-    const days = Math.floor(seconds / 86400);
-    const hours = Math.floor((seconds % 86400) / 3600);
-    const mins = Math.floor((seconds % 3600) / 60);
-    if (days > 0) return `${days}d ${hours}h`;
-    if (hours > 0) return `${hours}h ${mins}m`;
-    return `${mins}m`;
-  };
 
   return (
     <div className="page">
